Use 64-bits precision and add command line arguments
About 10% faster to generate a 1080p image using f64.

As for the CLI crate, I used `structopt`.
Shout-out to Pascal Hertleif for recommending it.

+++ src/vec3.rs
@@ -6,7 +6,7 @@ use std::ops::Mul;
 use std::ops::Neg;
 use std::ops::Sub;
 
-pub type Real = f32;
+pub type Real = f64;
 
 #[derive(Debug, Copy, Clone, PartialEq)]
 pub struct Vec3 {
